requirements.txt
aiocryptopay==0.4.8
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.4.0
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.13.0
arabic-reshaper==3.0.0
attrs==25.4.0
blinker==1.9.0
brotli==1.2.0
certifi==2023.11.17
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cssselect2==0.9.0
defusedxml==0.7.1
distro==1.9.0
Flask==3.1.3
fonttools==4.62.1
fpdf2==2.8.7
frozenlist==1.8.0
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
magic-filter==1.0.12
MarkupSafe==3.0.3
multidict==6.7.1
openai==2.30.0
pillow==12.1.1
playwright==1.58.0
propcache==0.4.1
pyaes==1.6.1
pyasn1==0.6.3
pycparser==3.0
pydantic==2.9.2
pydantic_core==2.23.4
pydyf==0.12.1
pyee==13.0.1
pyphen==0.17.2
python-bidi==0.6.7
python-telegram-bot==22.7
reportlab==4.4.10
requests==2.33.0
rsa==4.9.1
sniffio==1.3.1
StrEnum==0.4.15
Telethon==1.42.0
tinycss2==1.5.1
tinyhtml5==2.1.0
tqdm==4.67.3
typing_extensions==4.15.0
urllib3==2.6.3
weasyprint==68.1
webencodings==0.5.1
Werkzeug==3.1.7
yarl==1.23.0
zopfli==0.4.1
